Oppdatering av ekstern dekoratør
