+++ app/layout.tsx
@@ -3,16 +3,17 @@ import type { Metadata } from 'next';
 import '@navikt/arbeidsgiver-notifikasjon-widget/lib/cjs/index.css';
 import '@navikt/bedriftsmeny/lib/bedriftsmeny.css';
 import '@navikt/ds-css';
-import parse from 'html-react-parser';
+import Script from 'next/script';
 import './globals.css';
 
 import { Loader } from '@navikt/ds-react';
+import { fetchDecoratorReact } from '@navikt/nav-dekoratoren-moduler/ssr';
 import { NuqsAdapter } from 'nuqs/adapters/next';
 import { Suspense } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 import { ApplikasjonsContextProvider } from './ApplikasjonsContext';
-import { hentSsrDekoratør } from './components/dekoratør/dekoratør.server';
 import Header from './components/Header';
+import { hentMiljø, Miljø } from './util/miljø';
 
 export const metadata: Metadata = {
   title: 'Foreslåtte kandidater',
@@ -32,22 +33,56 @@ function RootSuspense({ children }: { children: React.ReactNode }) {
   );
 }
 
+export const byggBrødsmulesti = (miljø: Miljø) => {
+  if (miljø === Miljø.ProdGcp) {
+    return [
+      {
+        title: 'Min side – arbeidsgiver',
+        url: 'https://arbeidsgiver.nav.no/min-side-arbeidsgiver/',
+      },
+      {
+        title: 'Kandidater til dine stillinger',
+        url: 'https://arbeidsgiver.nav.no/kandidatliste/',
+      },
+    ];
+  } else {
+    return [
+      {
+        title: 'Min side – arbeidsgiver',
+        url: 'https://arbeidsgiver.intern.dev.nav.no/min-side-arbeidsgiver/',
+      },
+      {
+        title: 'Kandidater til dine stillinger',
+        url: 'https://presenterte-kandidater.intern.dev.nav.no/kandidatliste/',
+      },
+    ];
+  }
+};
+
 export default async function RootLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const dekoratør = await hentSsrDekoratør();
+  const Decorator = await fetchDecoratorReact({
+    env: hentMiljø() === Miljø.ProdGcp ? 'prod' : 'dev',
+    params: {
+      context: 'arbeidsgiver',
+      chatbot: false,
+      simple: false,
+      breadcrumbs: byggBrødsmulesti(hentMiljø()),
+    },
+  });
 
   return (
     <html lang='nb'>
       <head>
         <meta charSet='utf-8' />
         <meta name='viewport' content='width=device-width,initial-scale=1' />
-        {dekoratør && parse(dekoratør.styles)}
+        <Decorator.HeadAssets />
       </head>
       <body className='min-h-screen bg-[--a-gray-100]'>
-        {dekoratør && parse(dekoratør.header)}
+        <Decorator.Header />
         <RootSuspense>
           <div className='min-h-screen'>
             <div className='w-full border-b'>
@@ -58,7 +93,8 @@ export default async function RootLayout({
             </main>
           </div>
         </RootSuspense>
-        {dekoratør && parse(dekoratør.footer)}
+        <Decorator.Footer />
+        <Decorator.Scripts loader={Script} />
       </body>
     </html>
   );
